[HOTFIX] 최초 사진 등록 시 기존 사진 삭제 작업 안함

+++ src/main/java/hiff/hiff/behiff/domain/user/application/UserPhotoService.java
@@ -71,7 +71,9 @@ public class UserPhotoService {
 
     private void saveMainPhotoUrl(Long userId, String mainPhotoUrl) {
         User user = userCRUDService.findById(userId);
-        gcsService.deleteImage(user.getMainPhoto(), MAIN_PHOTO_FOLDER_NAME);
+        if(mainPhotoUrl != null) {
+            gcsService.deleteImage(user.getMainPhoto(), MAIN_PHOTO_FOLDER_NAME);
+        }
         user.updateMainPhoto(mainPhotoUrl);
     }
 }
